% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_entropy_units.R
\name{convert_entropy_units}
\alias{convert_entropy_units}
\title{Convert entropy units (based on )}
\usage{
convert_entropy_units(
  entropy,
  current_unit = c("log2", "ln"),
  new_unit = c("log2", "ln", "normalize"),
  alphabet_size
)
}
\arguments{
\item{entropy}{Entropy value (numeric)}

\item{current_unit}{Current unit of entropy (\code{"log2"} = bits, \code{"ln"} = nats)}

\item{new_unit}{Desired unit of entropy (\code{"log2"} for bits, \code{"ln"} for nats, and \code{"normalize"} for normalized entropy/evenness)}

\item{alphabet_size}{Size of the alphabet of the distribution whose entropy is in question (only needed if \code{new_unit == "normalize"})}
}
\value{
Entropy in desired units
}
\description{
Convert entropy units (based on )
}
\examples{
convert_entropy_units(entropy = 1.1, current_unit = "log2", new_unit = "ln")
}
