% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_ci_bootstrap_basic.R
\name{entropy_ci_bootstrap_basic}
\alias{entropy_ci_bootstrap_basic}
\title{Basic/pivot bootstrap confidence intervals for Shannon entropy
Confidence interval for Shannon entropy using the basic/pivot bootstrap method}
\usage{
entropy_ci_bootstrap_basic(
  bin_counts,
  pt_est_fct,
  B = 10^3,
  conf_level = 0.95,
  unit = c("log2", "ln", "normalize"),
  pre_calc_bootstraps = NULL
)
}
\arguments{
\item{bin_counts}{A numeric vector of observed bin counts.}

\item{pt_est_fct}{A function that computes the point estimate of entropy. The function must take \code{bin_counts} as its sole argument and return a scalar entropy estimate (in bits by default).}

\item{B}{A numeric value specifying the number of bootstrap resamples. Defaults to \code{1000}.}

\item{conf_level}{A numeric value representing the desired confidence level for the interval. Defaults to \code{0.95} (95\% confidence level).}

\item{unit}{A character string specifying the unit for entropy: \code{"log2"} for bits, \code{"ln"} for nats, or \code{"normalize"} for normalized entropy/evenness. Defaults to \code{"log2"}.}

\item{pre_calc_bootstraps}{Optional: A matrix of bootstrap bin counts, such as that produced by \code{rmultinom()}. Each column should represent the bin counts for one bootstrap sample. If provided, these pre-calculated bootstrap samples will be used instead of generating new ones.}
}
\value{
A list with the following components:
\item{ci}{A numeric vector containing the lower and upper bounds of the confidence interval for Shannon entropy.}
\item{pt_est}{The point estimate of Shannon entropy with the given point estimation function \code{pt_est_fct()}.}
\item{bootstrap_cdf}{The ECDF object representing the bootstrap distribution of entropy estimates.}
}
\description{
This function estimates a confidence interval for Shannon entropy using the basic/pivot bootstrap method.
}
\examples{
# Example usage with observed bin counts
entropy_ci_bootstrap_basic(bin_counts = c(4, 1, 5),
                           pt_est_fct = \(bin_counts) entropy_estimator_ML(bin_counts)$pt_est)

# Example with custom confidence level and number of bootstrap samples
entropy_ci_bootstrap_basic(bin_counts = c(4, 1, 5),
                           pt_est_fct = \(bin_counts) entropy_estimator_ML(bin_counts)$pt_est,
                           B = 2000,
                           conf_level = 0.90)

# Example with pre-calculated bootstrap samples
bin_counts <- c(4, 1, 5)
pre_boot <- rmultinom(1000, sum(bin_counts), bin_counts/sum(bin_counts))
entropy_ci_bootstrap_basic(bin_counts = bin_counts,
                           pt_est_fct = \(bin_counts) entropy_estimator_ML(bin_counts)$pt_est,
                           pre_calc_bootstraps = pre_boot)
}
