% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_estimator_Dirichlet.R
\name{entropy_estimator_Dirichlet}
\alias{entropy_estimator_Dirichlet}
\title{Bayesian entropy estimator with Dirichlet prior}
\usage{
entropy_estimator_Dirichlet(
  bin_counts,
  Dir_alpha = 1,
  unit = c("log2", "ln", "normalize")
)
}
\arguments{
\item{bin_counts}{Vector of observed bin counts}

\item{Dir_alpha}{Either (a) a numeric vector for the \eqn{\boldsymbol \alpha} parameter of the Dirichlet distriubiton or (b) a positive real \eqn{\alpha}, which is interpreted as the parameter \eqn{(\alpha , \dots, \alpha )}}

\item{unit}{Desired unit (\code{"log2"} for bits, \code{"ln"} for nats, and \code{"normalize"} for normalized entropy/evenness)}
}
\value{
A list containing three elements: entropy point estimate in the specified unit, an estimate of the probabilities (mean of \eqn{\mathbf p \sim \operatorname{Dirichlet}(\boldsymbol \alpha_{\text{posterior}})}), and the parameter \eqn{\boldsymbol \alpha_{\text{posterior}}} from the posterior Dirichlet distribution
}
\description{
Bayesian entropy estimator based on the prior \eqn{\mathbf p \sim \operatorname{Dirichlet}(\boldsymbol \alpha)}
}
\examples{
set.seed(123)
p <- c(.7, .15, .05, .05)
bin_counts <- rmultinom(1, 15, p)[,1]
print(bin_counts)
entropy_estimator_Dirichlet(bin_counts)
entropy_estimator_Dirichlet(bin_counts, Dir_alpha = c(4, 2, 1, 1), unit = "ln")
entropy_estimator_Dirichlet(bin_counts, unit = "normalize")
}
