% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_estimator_ML.R
\name{entropy_estimator_ML}
\alias{entropy_estimator_ML}
\title{Maximum likelihood entropy estimator (plugin estimator)}
\usage{
entropy_estimator_ML(bin_counts, unit = c("log2", "ln", "normalize"))
}
\arguments{
\item{bin_counts}{Vector of observed bin counts}

\item{unit}{Desired unit ("log2" for bits, "ln" for nats, and "normalize" for normalized entropy/evenness)}
}
\value{
List containing two elements: "pt_est" in the specified unit; "prob_est_ML" the ML point estimates of bin probabilities
}
\description{
Maximum likelihood entropy estimator (plugin estimator)
}
\examples{
set.seed(123)
p <- c(.7, .15, .05, .05)
bin_counts <- rmultinom(1, 15, p)[,1]
print(bin_counts)
entropy_estimator_ML(bin_counts)
entropy_estimator_ML(bin_counts, unit = "ln")
entropy_estimator_ML(bin_counts, unit = "normalize")
}
