% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_estimator_MM.R
\name{entropy_estimator_MM}
\alias{entropy_estimator_MM}
\title{Miller-Madow entropy estimator}
\usage{
entropy_estimator_MM(bin_counts, unit = c("log2", "ln", "normalize"))
}
\arguments{
\item{bin_counts}{Vector of observed bin counts}

\item{unit}{Desired unit (\code{"log2"} for bits, \code{"ln"} for nats, and \code{"normalize"} for normalized entropy/evenness)}
}
\value{
List containing one element: the entropy point estimate (\code{"pt_est"}) in the specified unit
}
\description{
Miller-Madow entropy estimator
}
\examples{
set.seed(123)
p <- c(.7, .15, .05, .05)
bin_counts <- rmultinom(1, 15, p)[,1]
print(bin_counts)
entropy_estimator_MM(bin_counts)
entropy_estimator_MM(bin_counts, unit = "ln")
entropy_estimator_MM(bin_counts, unit = "normalize")
}
