% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_estimator_SG.R
\name{entropy_estimator_SG}
\alias{entropy_estimator_SG}
\title{Bayesian entropy estimator with the Schürmann-Grassberger Dirichlet prior}
\usage{
entropy_estimator_SG(
  bin_counts,
  est_type = c("mean_entropy_of_p", "entropy_of_mean_p"),
  unit = c("log2", "ln", "normalize")
)
}
\arguments{
\item{bin_counts}{Vector of observed bin counts}

\item{est_type}{Type of entropy estimator to use: either \code{"mean_entropy_of_p"} for \eqn{\mathbb E[H(p)]} or \code{"entropy_of_mean_p"} for \eqn{H(\mathbb E[p])},
where the expectation is taken over the posterior Dirichlet distribution of \eqn{p}}

\item{unit}{Desired unit (\code{"log2"} for bits, \code{"ln"} for nats, and \code{"normalize"} for normalized entropy/evenness)}
}
\value{
A list containing three elements: entropy point estimate in the specified unit, an estimate of the probabilities (mean of \eqn{\mathbf p \sim \operatorname{Dirichlet}(\boldsymbol \alpha_{\text{posterior}})}), and the parameter \eqn{\boldsymbol \alpha_{\text{posterior}}} from the posterior Dirichlet distribution
}
\description{
Bayesian entropy estimator based on the Schürmann-Grassberger prior \eqn{\mathbf p \sim \operatorname{Dirichlet}\big((1/K, \dots, 1/K)\big)}, where \eqn{k \hat= } alphabet size
}
\examples{
set.seed(123)
p <- c(.7, .15, .05, .05)
bin_counts <- rmultinom(1, 15, p)[,1]
print(bin_counts)
entropy_estimator_SG(bin_counts)
entropy_estimator_SG(bin_counts, unit = "ln")
entropy_estimator_SG(bin_counts, unit = "normalize")
}
