% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_lists_with_names.R
\name{merge_lists_with_names}
\alias{merge_lists_with_names}
\title{Merge Two Named Lists Without Duplicates}
\usage{
merge_lists_with_names(l1, l2)
}
\arguments{
\item{l1}{A named list. The first input list to be merged.}

\item{l2}{A named list. The second input list to be merged.}
}
\value{
A named list containing all unique elements from both input lists. If
duplicate names exist, only the first occurrence is retained.
}
\description{
This function merges two named lists while ensuring that elements with the
same name appear only once in the resulting list. Only the first occurrence
of each name is kept.
}
\examples{
list1 <- list(a = list(1, 2, 3), b = list(4, 5))
list2 <- list(b = list(5, 6), c = list(7, 8))
merge_lists_with_names(list1, list2)
# Returns:
# $a
# [[1]]
# [1] 1
#
# [[2]]
# [1] 2
#
# [[3]]
# [1] 3
#
# $b
# [[1]]
# [1] 4
#
# [[2]]
# [1] 5
#
# $c
# [[1]]
# [1] 7
#
# [[2]]
# [1] 8
}
