% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon entropy estimation}
\usage{
shannon_entropy(
  bin_counts,
  total_bins_known = TRUE,
  point_estimator.method = c("ML", "MM", "HS"),
  confidence_interval.method = c("Wald", "Asym.Norm_quadratic", "pct_boot",
    "multinom_ci_based_Wallis", "multinom_ci_based_Erhard"),
  confidence_level = 0.95,
  B = 10^3,
  normalize = TRUE
)
}
\arguments{
\item{bin_counts}{Vector of bin counts}

\item{total_bins_known}{Logical vector indicating if number of bins known}

\item{point_estimator.method}{Point estimator}

\item{confidence_interval.method}{Confidence interval estimator}

\item{confidence_level}{Confidence level (default = 0.95)}

\item{B}{Number of bootstrap re-samplings (if applicable)}

\item{normalize}{Logical vector for normalization (TRUE -> normalized entropy)}
}
\value{
Point estimate and confidence interval
}
\description{
Shannon entropy estimation
}
\examples{
shannon_entropy(c(100,9,0), TRUE, "ML", "Wald", normalize = TRUE)
}
