% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_bin_counts.R
\name{valid_bin_counts}
\alias{valid_bin_counts}
\title{Checks validity of \code{bin_counts} parameter}
\usage{
valid_bin_counts(bin_counts)
}
\arguments{
\item{bin_counts}{Parameter bin_counts to check}
}
\value{
\code{NULL} if valid, stops function with an error message otherwise
}
\description{
Checks if \code{bin_counts} parameter is a vector of non-negative integers, not all zero.
}
\examples{
valid_bin_counts(-1:10)
valid_bin_counts(c(3, 3.1, 4))
valid_bin_counts(1:5)
}
